Simplify extending cluster handle implementations
